requirements.txt
apply-defaults==0.1.1
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
click==6.7
colorama==0.4.1
dataclasses==0.6
Django==2.2.3
django-mongodb-engine==0.6.0
django-redis==4.10.0
djangorestframework==3.10.1
djangotoolbox==1.8.0
djongo==1.2.33
et-xmlfile==1.1.0
excel2json-3==0.1.6
Flask==1.1.1
Flask-MongoAlchemy==0.7.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
json-rpc==1.12.1
jsonrpc==1.2
jsonrpcclient==3.3.3
jsonrpcserver==4.0.4
jsonschema==3.0.2
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matlabengineforpython===R2020a
matlabruntimeforpython===R2020a
mccabe==0.6.1
MongoAlchemy==0.19
mypy==0.720
mypy-extensions==0.4.1
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
nltk==3.4.4
numpy==1.18.4
opencv-python==4.2.0.34
openpyxl==3.0.7
panda==0.3.1
pandas==1.2.4
plumbum==1.6.7
pycodestyle==2.5.0
pylint==2.3.1
pymongo==3.8.0
pyrsistent==0.15.4
python-dateutil==2.8.1
pytz==2019.1
redis==3.2.1
requests==2.22.0
rpyc==4.1.1
six==1.12.0
SQLAlchemy==1.3.6
sqlparse==0.2.4
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
virtualenv==16.7.2
Werkzeug==0.15.5
wrapt==1.11.2
xlrd==2.0.1
